requirements.txt
google-cloud-aiplatform
google-cloud-storage
langchain
langchain-google-vertexai
pypdf
